refactor(ui): split the selection handlers and key popover
Both files had grown past the guidelines in file-splitting.md — the
selection hook to 723 lines against 600, the popover to 563 against 500 —
after the Auto Move follow-along landed.

What moved out is what could leave without taking a shared invariant with
it. The next-key lookup and anchor measurement are pure functions of the
walk order and the DOM, so they became a plain module. Undo/redo is one
concern with one entry point, so it became its own hook. The popover's
wrapper mode and its pending-confirm flow moved together, because the mode
decides what shape a confirm takes and splitting them would have spread
that decision across two files.

What stayed is the part that must not be split: the four selection states
and the epoch guarding them are a single invariant — an async write must
never clobber a selection that has since been replaced. Handing that
control around would have undone the consolidation it just took to make
that invariant structural. The history hook therefore receives two narrow
functions, never the ref, so the epoch still has exactly one place that
can advance it.

No behaviour change: the suite reports the same counts before and after,
and no test needed rewriting.
