+++ src/renderer/components/editors/__tests__/useKeymapSelectionHandlers.popoverAdvance.test.ts
@@ -4,7 +4,8 @@
 import { useRef } from 'react'
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { renderHook, act } from '@testing-library/react'
-import { useKeymapSelectionHandlers, nextAdvanceKey } from '../useKeymapSelectionHandlers'
+import { useKeymapSelectionHandlers } from '../useKeymapSelectionHandlers'
+import { nextAdvanceKey } from '../keymap-auto-advance'
 import { useKeymapMultiSelect } from '../useKeymapMultiSelect'
 import { useKeymapHistory } from '../useKeymapHistory'
 import type { UseKeymapSelectionOptions } from '../useKeymapSelectionHandlers'
