+++ src/renderer/components/editors/useKeymapSelectionHandlers.ts
@@ -13,48 +13,10 @@ import type { MacroAction } from '../../../preload/macro'
 import { hasModifierKey } from './KeyboardPane'
 import type { PopoverState } from './keymap-editor-types'
 import type { UseKeymapMultiSelectReturn } from './useKeymapMultiSelect'
-import type { UseKeymapHistoryReturn, SingleHistoryEntry, HistoryEntry } from './useKeymapHistory'
-import { sortKeysByViewMatrix, type ViewMatrixKeyRef } from './view-matrix'
-
-/**
- * Given the Auto Move walk order (already layout-option/decal/encoder
- * filtered and view-matrix sorted — see `advancableKeys` below) and a
- * starting position, returns the next key in the walk, or `null` when
- * `from` isn't in the list at all or is already last. Pure — callers
- * (both the left-panel picker and the popover follow-along) gate
- * `autoAdvance` themselves before calling this.
- */
-export function nextAdvanceKey<K extends ViewMatrixKeyRef>(
-  advancableKeys: readonly K[],
-  from: { row: number; col: number },
-): { row: number; col: number } | null {
-  const idx = advancableKeys.findIndex((k) => k.row === from.row && k.col === from.col)
-  if (idx < 0 || idx >= advancableKeys.length - 1) return null
-  const next = advancableKeys[idx + 1]
-  return { row: next.row, col: next.col }
-}
-
-/**
- * Finds the on-screen rect for the key at physical `(row, col)`, scoped to
- * `container` — never `document`. The Keyboard tab's hidden layout picker
- * renders its own elements with the same `data-key-pos` attribute
- * (`KeyWidget.tsx`), so an unscoped lookup can resolve to the wrong pane's
- * key; `container` must be the primary keymap pane's own content ref.
- *
- * Scrolls the element into view first — `behavior: 'instant'` is required
- * (not the default 'auto') so a key currently scrolled out of view is
- * measured at its final position instead of mid smooth-scroll animation.
- * Returns `null` when no matching element exists (e.g. a key hidden by the
- * current layout option, or the container isn't mounted) — callers must
- * not advance the popover in that case.
- */
-function getKeyAnchorRect(container: HTMLElement | null, row: number, col: number): DOMRect | null {
-  if (!container) return null
-  const el = container.querySelector(`[data-key-pos="${row},${col}"]`)
-  if (!el) return null
-  el.scrollIntoView?.({ block: 'nearest', inline: 'nearest', behavior: 'instant' })
-  return el.getBoundingClientRect()
-}
+import type { UseKeymapHistoryReturn, SingleHistoryEntry } from './useKeymapHistory'
+import { sortKeysByViewMatrix } from './view-matrix'
+import { nextAdvanceKey, getKeyAnchorRect } from './keymap-auto-advance'
+import { useKeymapHistoryActions } from './use-keymap-history-actions'
 
 /** Partial write to the selection-state quartet (`selectedKey` /
  *  `selectedMaskPart` / `selectedEncoder` / `popoverState`) that
@@ -66,19 +28,6 @@ interface SelectionPatch {
   selectedMaskPart?: boolean
   selectedEncoder?: { idx: number; dir: 0 | 1 } | null
   popoverState?: PopoverState | null | ((prev: PopoverState | null) => PopoverState | null)
-}
-
-/** Match a history entry against the current popover position, returning the keycode if matched. */
-function matchPopoverEntry(
-  popoverState: PopoverState | null,
-  entry: HistoryEntry | null,
-  currentLayer: number,
-  field: 'oldKeycode' | 'newKeycode',
-): number | undefined {
-  if (!popoverState || !entry || entry.kind === 'batch') return undefined
-  if (popoverState.kind === 'key' && entry.kind === 'key' && entry.layer === currentLayer && entry.row === popoverState.row && entry.col === popoverState.col) return entry[field]
-  if (popoverState.kind === 'encoder' && entry.kind === 'encoder' && entry.layer === currentLayer && entry.idx === popoverState.idx && entry.dir === popoverState.dir) return entry[field]
-  return undefined
 }
 
 export interface UseKeymapSelectionOptions {
@@ -108,15 +57,10 @@ export interface UseKeymapSelectionOptions {
   multiSelect: UseKeymapMultiSelectReturn
   // History
   history: UseKeymapHistoryReturn
-  /** Fires the "flash" visual (see `useKeyFlash`) for the positions an
-   *  undo/redo just touched. Contract: called only after ALL of that
-   *  entry's device writes have succeeded AND the history stack has been
-   *  committed (`history.undo()` / `history.redo()`) — never on a failed
-   *  apply, and never before the commit. An exception thrown by this
-   *  callback must not retroactively mark the undo/redo as failed, so it
-   *  is invoked after `handleUndo`/`handleRedo`'s own try/finally has
-   *  already run to completion. Receives the normalized entry list
-   *  (`entry.kind === 'batch' ? entry.entries : [entry]`). */
+  /** Forwarded to `useKeymapHistoryActions` — see the contract on
+   *  `onHistoryApplied` in `use-keymap-history-actions.ts`'s
+   *  `UseKeymapHistoryActionsOptions` (that hook's `runHistoryStep` is
+   *  what actually invokes it). */
   onHistoryApplied?: (entries: SingleHistoryEntry[]) => void
   // TD/Macro
   tapDanceEntries?: TapDanceEntry[]
@@ -575,112 +519,23 @@ export function useKeymapSelectionHandlers({
     }
   }, [popoverState, currentLayer, keymap, encoderLayout, onSetKey, onSetEncoder, history])
 
-  // --- History-derived popover undo ---
-  const popoverUndoKeycode = useMemo(
-    () => matchPopoverEntry(popoverState, history.peekUndo, currentLayer, 'oldKeycode'),
-    [popoverState, currentLayer, history.peekUndo],
-  )
-
-  // --- Undo / redo ---
-  const applyHistoryEntry = useCallback(async (entry: HistoryEntry, isUndo: boolean) => {
-    if (entry.kind === 'batch') {
-      const items = isUndo ? [...entry.entries].reverse() : entry.entries
-      const keyEntries: BulkKeyEntry[] = []
-      const encoderOps: { layer: number; idx: number; dir: number; code: number }[] = []
-      for (const e of items) {
-        const code = isUndo ? e.oldKeycode : e.newKeycode
-        if (e.kind === 'key') keyEntries.push({ layer: e.layer, row: e.row, col: e.col, keycode: code })
-        else encoderOps.push({ layer: e.layer, idx: e.idx, dir: e.dir, code })
-      }
-      if (keyEntries.length > 0) await onSetKeysBulk(keyEntries)
-      for (const op of encoderOps) await onSetEncoder(op.layer, op.idx, op.dir, op.code)
-    } else {
-      const code = isUndo ? entry.oldKeycode : entry.newKeycode
-      if (entry.kind === 'key') await onSetKey(entry.layer, entry.row, entry.col, code)
-      else await onSetEncoder(entry.layer, entry.idx, entry.dir, code)
-    }
-  }, [onSetKey, onSetKeysBulk, onSetEncoder])
-
-  // In-flight guard to prevent concurrent undo/redo
-  const undoRedoInFlightRef = useRef(false)
-
-  // Undo and redo differ only in direction (which stack to peek/commit,
-  // and which side of the entry `applyHistoryEntry` restores) — shared here
-  // instead of duplicating the guard/apply/commit/notify sequence twice.
-  const runHistoryStep = useCallback(async (isUndo: boolean) => {
-    if (undoRedoInFlightRef.current) return
-    const entry = isUndo ? history.peekUndo : history.peekRedo
-    if (!entry) return
-    undoRedoInFlightRef.current = true
-    // Snapshot the epoch before the device writes below — which, for a
-    // batch entry, run as a sequence of awaited HID calls and can take
-    // long enough for the user to open a different popover in the
-    // meantime. `undoRedoInFlightRef` only blocks a second concurrent
-    // undo/redo; it does nothing to stop that other popover from opening.
-    // Re-checking the epoch here before closing the popover keeps this
-    // consistent with every other popover write handler in this file:
-    // a stale completion must never clobber something newer.
-    const epoch = popoverAdvanceEpochRef.current
-    try {
-      await applyHistoryEntry(entry, isUndo)
-      // Commit only after successful apply.
-      if (isUndo) history.undo()
-      else history.redo()
-    } finally { undoRedoInFlightRef.current = false }
+  // --- Undo / redo (epoch read + conditional close handed over as a small
+  // controller — see the option doc on `useKeymapHistoryActions` for why
+  // the ref itself is never passed across the boundary) ---
+  const getPopoverEpoch = useCallback((): number => popoverAdvanceEpochRef.current, [])
+  const closePopoverIfEpochMatches = useCallback((epoch: number) => {
     if (popoverAdvanceEpochRef.current === epoch) applySelectionChange({ popoverState: null })
-    // Fire outside the try/finally above: a throw from `applyHistoryEntry`
-    // or the commit call propagates out of the `try` (after `finally`
-    // resets the in-flight guard) and skips everything below, so reaching
-    // this line already guarantees the apply + commit succeeded — no flag
-    // needed to gate it. Placement after the commit is what guarantees
-    // `onHistoryApplied` can no longer un-commit the undo/redo or leave the
-    // in-flight guard stuck. The flash it triggers is purely cosmetic, so a
-    // throw from the callback itself is swallowed here rather than
-    // rejecting `runHistoryStep`'s promise — the undo/redo already
-    // succeeded and must not be reported as failed just because the flash
-    // visual couldn't be shown.
-    try {
-      onHistoryApplied?.(entry.kind === 'batch' ? entry.entries : [entry])
-    } catch {
-      // Intentionally ignored — see comment above.
-    }
-  }, [history, applyHistoryEntry, onHistoryApplied, applySelectionChange])
+  }, [applySelectionChange])
 
-  const handleUndo = useCallback(() => runHistoryStep(true), [runHistoryStep])
-  const handleRedo = useCallback(() => runHistoryStep(false), [runHistoryStep])
-
-  const handlePopoverUndo = useCallback(() => {
-    if (popoverUndoKeycode == null) return
-    void handleUndo()
-  }, [popoverUndoKeycode, handleUndo])
-
-  // --- History-derived popover redo (top-only) ---
-  const popoverRedoKeycode = useMemo(
-    () => matchPopoverEntry(popoverState, history.peekRedo, currentLayer, 'newKeycode'),
-    [popoverState, currentLayer, history.peekRedo],
-  )
-
-  const handlePopoverRedo = useCallback(() => {
-    if (popoverRedoKeycode == null) return
-    void handleRedo()
-  }, [popoverRedoKeycode, handleRedo])
-
-  // --- Keyboard shortcuts for undo/redo ---
-  useEffect(() => {
-    function onKeyDown(e: KeyboardEvent) {
-      const target = e.target as Element | null
-      if (target instanceof HTMLInputElement || target instanceof HTMLTextAreaElement) return
-      if (target?.closest?.('[contenteditable]')) return
-      const mod = e.ctrlKey || e.metaKey
-      if (!mod) return
-      const key = e.key.toLowerCase()
-      if (key === 'z' && !e.shiftKey) { e.preventDefault(); void handleUndo(); return }
-      if (key === 'z' && e.shiftKey) { e.preventDefault(); void handleRedo(); return }
-      if (key === 'y' && !e.shiftKey) { e.preventDefault(); void handleRedo(); return }
-    }
-    window.addEventListener('keydown', onKeyDown)
-    return () => window.removeEventListener('keydown', onKeyDown)
-  }, [handleUndo, handleRedo])
+  const {
+    popoverUndoKeycode, handlePopoverUndo,
+    popoverRedoKeycode, handlePopoverRedo,
+    handleUndo, handleRedo,
+  } = useKeymapHistoryActions({
+    history, popoverState, currentLayer,
+    onSetKey, onSetKeysBulk, onSetEncoder,
+    onHistoryApplied, getPopoverEpoch, closePopoverIfEpochMatches,
+  })
 
   // --- TD/Macro modal handlers ---
   const handleTdModalSave = useCallback(async (idx: number, entry: TapDanceEntry) => {
